Construct outdata class

The outdata class define a standard output format for analysis and reporting.

Usage

outdata(
  meta,
  population,
  observation,
  parameter,
  n,
  order,
  group,
  reference_group,
  ...
)

Arguments

meta

a meta data created by metalite.

population

a character value of population term name. The term name is used as key to link information.

observation

a character value of observation term name. The term name is used as key to link information.

parameter

a character value of parameter term name. The term name is used as key to link information.

n

a dataframe for number of subjects in each criteria

order

a numeric vector of row display order

group

a character vector of group variable names in an ADaM dataset.

reference_group

a numeric value to indicate reference group in levels of group.

...

additional variables save to outdata.

Value

A list with class outdata. Components of the list are either quosures or constants.

Details

The implementation idea mimic ggplot2::aes.